Add package for memory storage of database runtime
Database configuration data for all namespaces, databases, logins, tokens, scopes, tables, fields, indexes should be stored in memory for quick access when needed without querying the KV store.

It should be kept in sync with the master data in the unerlying KV store, but it can be used for propagating the database configuration across the cluster.