Integrate `surrealdb-rs` into `surrealdb`
Since `surrealdb` can be embedded into `surrealdb-rs`, it makes sense to
combine these two crates so we can expose a single API that works for both
embedded and remote databases. This eliminates confusion as to which crate
to use.

Previously the answer to that question would be “use `surrealdb` if you want
an embedded database or use `surrealdb-rs` if you want to connect to a remote
one”. Now the answer to this question is more nuanced.