Fix bug that prevented the fuzzers from using dictionaries properly
The original fuzzers where using the arbitrary::Arbitrary derivations
to generate a Vec of surql commands to pass to the database. However
when elements of the dictionary where inserted into the fuzzing
stream the arbitrary::Arbitrary derived generated was mangling the
input before it got the database. To fix this we've just converted
the Vec to a single string, which doesn't mangle the input. Then
we split on ';' to generated an equivelent Vec of commands.

Closes: #2027