Add boolean/logical array functions (currently `boolean_and`, `boolean_or`, and `boolean_xor`) and tests.
Note that for now, the `boolean_or` test does not pass for bizzare reasons. Due to the fact that the
same test query set works in Surrealist, I suspect that this is a bug in the Datastore::execute function
or something along the ways. because of this, I'll ignore the test for now and leave it to the massive
cleanup of failing tests that will have to come soon.