Implement execution of async events
The test suite passed for the non-async case, I will add async event tests and debug next.

It tokio::spaws the event execution future and drops the join handle, such that it keeps running as a background task. It is wrapped in an independent transaction (Context) and executed with a fresh stack (Stk).

Async events are not supported in WASM for now. Added a syntax error when doing DEFINE EVENT ASYNC in WASM. If an async event is executed in WASM regardless, it will be run synchronously like non-async events with a tracing::warn.